refactor: Relocate theme files and update MainActivity package
