+++ app/src/main/java/com/london/novix/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.london.novix
+package com.london.novix.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.london.novix.ui.theme.NovixTheme
+import com.london.novix.presentation.designSystem.theme.NovixTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
